feat(webhook): include signed header for webhook response
https://bugzilla.redhat.com/show_bug.cgi?id=1213630
https://github.com/zanata/zanata-server/pull/904

Squashed commit of the following:

commit 872def0462046cd57d30dc6c895271f18b647263
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 16 11:50:15 2015 +1000

    Add pseudocode in doc for verify webhook request

commit 52366c98ca6e423cafad8d363d751fce4442b1a9
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 16 11:16:56 2015 +1000

    Update doc with double hash information

commit f7f851a5a761208ae2fc229a8ce478567ca28cef
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 15 13:41:09 2015 +1000

    Fix null and empty secret key for webhook

commit 78f20f5246862c09513c7946c8d0fcb9cb519b4d
Author: Alex Eng <aeng@redhat.com>
Date:   Wed Jul 15 06:46:22 2015 +1000

    Update documentation

commit fbb4ef5476dcd6549ed91b7cd28001689008b7d4
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 14 16:36:27 2015 +1000

    Fix functional test

commit a98ee2dddd677532c9ba15c21ba2b6d22a8ce57a
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 14 13:13:41 2015 +1000

    Move secret to each webhook url

commit beb37070b098f1b19a8f8e2708566eaab6c26a2e
Author: Alex Eng <aeng@redhat.com>
Date:   Tue Jul 14 10:58:39 2015 +1000

    rename server url to callbackURL, add unit test for signHeader

commit 2c125857e2cdff7221e37350ad881eb8bdd9cda6
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 13 13:35:26 2015 +1000

    Fix enforcer issue with package dependency

commit e184c8f39201b488ac286ba50b42cfb46b4852c0
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 13 12:35:13 2015 +1000

    Rename test to integration test, update webhook field to be secret field

commit 16e14b137ab8c0ac5be128b9dc3ce34f6268ac4d
Author: Alex Eng <aeng@redhat.com>
Date:   Mon Jul 13 10:27:37 2015 +1000

    Exlude banned dependency

commit 7a421d74c9a34d632d5fc1ae958356d5123bdd8f
Author: Alex Eng <aeng@redhat.com>
Date:   Sat Jul 11 17:04:40 2015 +1000

    Update docs, update test and rename webhookKey to webhookSecret

commit 92f1e0f2873eaeae6f5e00d80476f5403ae8e976
Author: Alex Eng <aeng@redhat.com>
Date:   Fri Jul 10 16:18:53 2015 +1000

    Update button loader

commit 8a091c277b4d6fa1c1ca2f24f8331f0a7a700f84
Author: Alex Eng <aeng@redhat.com>
Date:   Fri Jul 10 15:35:12 2015 +1000

    feat(webhook): add zanata header for webhook

    https://bugzilla.redhat.com/show_bug.cgi?id=1213630

commit 434ce8429824b59be93392a1d3e8081003c9e87f
Merge: d6e68c5 fd95095
Author: Alex Eng <aeng@redhat.com>
Date:   Fri Jul 10 08:25:17 2015 +1000

    Merge branch 'integration/master' into rhbz1213630-webhook

commit d6e68c5bbe00db0a21549877202ad9beb3a6f80e
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 9 16:08:58 2015 +1000

    Add hmacsha1 header if secret key avaiQlable

commit d4780c136b8eb153565d1195375688c61c5b066a
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 9 14:17:02 2015 +1000

    Add document url into webhook post data

commit 2144c4f2de690d726d80e5ce42d8fe65828410fc
Author: Alex Eng <aeng@redhat.com>
Date:   Thu Jul 9 10:50:36 2015 +1000

    Add secret key for webhook in HProject

+++ zanata-war/src/main/java/org/zanata/dao/WebHookDAO.java
@@ -10,7 +10,7 @@ import org.zanata.model.WebHook;
 @Name("webHookDAO")
 @AutoCreate
 @Scope(ScopeType.STATELESS)
-public class WebHookDAO extends AbstractDAOImpl<WebHook, Integer> {
+public class WebHookDAO extends AbstractDAOImpl<WebHook, Long> {
 
     public WebHookDAO() {
         super(WebHook.class);
